Restrict stdlib embed macros to single source file
