+++ source/slang-stdlib/slang-embedded-stdlib.cpp
@@ -1,18 +1,8 @@
-// slang-stdlib-api.cpp
-
 #include "../core/slang-basic.h"
 #include "../core/slang-array-view.h"
-
 #include "../core/slang-blob.h"
 
-#ifdef SLANG_WITHOUT_EMBEDDED_STD_LIB
-
-SLANG_API ISlangBlob* slang_getEmbeddedStdLib()
-{
-    return nullptr;
-}
-
-#else
+#ifdef SLANG_EMBED_STDLIB
 
 static const uint8_t g_stdLib[] =
 {
@@ -24,6 +14,13 @@ static Slang::StaticBlob g_stdLibBlob((const void*)g_stdLib, sizeof(g_stdLib));
 SLANG_API ISlangBlob* slang_getEmbeddedStdLib()
 {
     return &g_stdLibBlob;
+}
+
+#else
+
+SLANG_API ISlangBlob* slang_getEmbeddedStdLib()
+{
+    return nullptr;
 }
 
 #endif
